Removed Conservation of Momentum Physcis
With the current setup I have, it would be impossible without involving other forces aside from kinetic. The reason behind this is because when stars collide, their collective energy isn't just expressed kinetically, but also visually as it produces an enormous amount of light. This is possible to add in later, however I'm just going to remove this for now to make it functional.